Gradient... bars... with sdl?
This might be the most useless fking shit I've done in a long, long time. Not only did I spend 6 hours on this (well not really, adhd clicking links in internet and getting a bunch of ideas for next), but midway I also realized I will need to revert this change.

Only comitting so I don't need to redo gradient later if I need it.